build: Setup user storage for tests
Add "make storage" target for configuring user storage for storage
tests. We run "make storage" in Travis CI.

In oVirt CI we also need to tear down the storage when exiting,
otherwise mock fail to remove the chroot.

Developer can run "make storage" manually to configure storage for local
tests. This is interesting only to people modifying storage code, so it
is not a hard dependency.

Change-Id: Ife44a1adda2151467d7ece76c599d6540a08a4da
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>