spec: Don't require libvirt and netaddr Python packages in configure
Those packages are needed only to run tests, not to e.g. build rpm's.
They are not sufficient for running the tests and they make rpm builds
failing.

We can approach the problem in two ways:

- To add the packages to BuildRequires.
- To remove the requirement in configure.

Adding useless dependencies to BuildRequires is not very good, so we
choose the latter option in this patch.  Dependencies for running the
tests should be addressed in a different way, in future patches.

Change-Id: If663aaf866e69ac6f6ee1f41efea3926f3ec59ff
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>