storage: add _blk suffix to local variables contaning size in blocks
As the first step to replace blocks with bytes, rename variables
containing size in blocks to have _blk suffix. This will make easy
identification of all affected variables later on.

This is the first patch of a patch set which aims to rename all
variables, functions and function parameters which use size in blocks
to have _blk suffix. The rename plan is as follows:
* rename local variables (this patch)
* refactor expressions into local variables
* rename functions
* rename function parameters

Change-Id: I0e57a07f2774e6a71b625987569fc489d5ce6cc7
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>