tool: Simplify python 3 support
This basically reverts commit b432a97dea41 (py3: Fix bytes x strings in
_lvmetad_configured) and fix it in a simpler and more maintaiable way by
decoding systemctl output.

The general rule to support unicode in a portable way is to use unicode
internally and utf-8 externally. When we read commands output that
always contain text, we should decode the output instead of creating a
mess of strings and bytes.

Change-Id: Ifd2a56a6892953e82c83538d0bbaf7afe4ea5029
Signed-off-by: Nir Soffer <nsoffer@redhat.com>