Fix dependency on sos on power platform
We were fixing sos dependency and removing VDSM sos plugin in [1], but
the patch had issue on power platforms. Power8 is going to be bumped to
RHEL 7.7, but unfortunately Power9 stays on RHEL 7.6-ALT.

This is oVirt 4.3 patch only, because in 4.4 all hosts is going to
require RHEL 8 on all supported platforms.

[1] https://gerrit.ovirt.org/100623

Change-Id: I76661c1520788e62f0d397f33942361ce64a6ef9
Bug-Url: https://bugzilla.redhat.com/1714509
Signed-off-by: Martin Perina <mperina@redhat.com>