storage: add _blk suffix to function arguments
Add _blk suffix to function arguments which are used to pass in size
in block. This patch is the last step in renaming functions and
variables which uses size in blocks. After this patch all variables
and functions which contains or returns values in blocks should have
_blk or Blk suffix. The only exceptions are public APIs and metadata
size.

In tests, we remove the default argument not to have it change again
when we switch to bytes.

Change-Id: Ie9d5d0a48b43c978e32985548ce751d47040ac3b
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>