stomp: Fix 'COMMANDS' tuple
A subtle bug has been hiding in the definition of the 'COMMANDS' tuple.
The values of the tuple are in fact attribute names of the 'Command'
class, *not* their values. Clients of 'COMMANDS' expect the latter:

lib/yajsonrpc/stompserver.py:402:

 REQUIRED_SIZE = max(len(s) for s in stomp.COMMANDS)

lib/yajsonrpc/stompserver.py:409:

 return data.startswith(stomp.COMMANDS)

The only thing that prevented the bug from manifesting is the fact that
for all the commands the attribute name is the same as the value.
Introducing a new command like:

 class Command(object):
     ...
     NEW_COMMAND = "SIGNATURE"

would break our code.

Change-Id: I9e26dcfc7cced1d67844f268721c66485b089906
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>