constants: Use diskType as a string value
Patch https://gerrit.ovirt.org/#/c/101068/ changed the diskType to int
since engine 4.1 sends diskType=2.
On the other hand, a cluster with 4.1 compatibility version requires
a string value "2" as it's diskType, since it sends diskType="2".

This patch converts diskType to a string to mitigate the inconsistent
behavior.

This is a patch for 4.3 only, since master version doesn't support
4.1 engine and cluster compatibility.

Change-Id: I7613f7ca524c36f2e0db99e634978f8ad9b84d7d
Bug-Url: https://bugzilla.redhat.com/1725390
Signed-off-by: Shani Leviim <sleviim@redhat.com>