tests: Test volume lease on block storage
Add missing tests for:
- Getting lease info
- Creating volume lease when creating a volume

Fix the exiting tests to use the volume metadata slot to calculate the
lease offset instead of hard-coded value that was correct only for
domain version 4.

Enable the tests for all supported domain versions, previously only
version 4 was tested.

Because creating a storage domain and volume is very slow add more
tests in the volume_life_cycle test.

Change-Id: I1b0481f61fac625524c8b0c6cbd78e05bc168d72
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>