blockSD: Use alignment to calculate leases and xleases volume size
Use domain alignment to calculate the size of the leases and xlease
volumes, instead of hard-coded size (assuming alignment 1m).

On file storage we create the leases and xleases as empty files and
writing to these files allocate space as needed, so no change is
required to support 4k storage.

With this change we can create the special volumes with 4k storage and
alignment != 1m and use volume and external leases.

More work is needed to read and write domain and volume metadata with 4k
storage.

Change-Id: Icb2ffb0af04463b1c7535042124799241685cefd
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>