net: Fix error if net removed during stats request
When calling a net stats report, a list of interfaces is
generated, and statistics are gathered in turn.
When an interface is removed, the statistics call might still
be running - containing the removed interface as active.
This causes a "device is not present in the system" error.

Fix by catching this specific exception and ignoring it
only during gathering of statistics in the report.

Change-Id: I06a3f51cfd923267bc72ccf113babbd64394336a
Bug-Url: https://bugzilla.redhat.com/1723668
Signed-off-by: Bell Levin <blevin@redhat.com>
(cherry picked from commit 5604d038d85d692087ca037819fda0d066c1eb5a)