Host api: write message to vdsm.log
In automatic testing it is very helpful for debugging to delineate the
start of each test in the log or to add a log message at any desirable
point in the flow.

- Engine already provides a REST API endpoint to add events with user
  messages to engine.log. This API has been used in ovirt-system-tests
  network-suite to automatically create a log message when a test starts
  (see gerrit.ovirt.org/#/c/100295).
- ovirt-system-tests/network-suite/ovirtlib can also be used to log
  messages on the fly (see gerrit.ovirt.org/#/c/100295).
- A bash script to send an event with a log message to engine has been
  added to ovirt-engine/contrib (see gerrit.ovirt.org/#/c/101482).

The purpose of this patch is to introduce a similar capability to
vdsm which will be helpful for synchronized delineation across
engine.log and vdsm.log.

Change-Id: I63939b31d18d09f842faa7cb8cfb6d10d798bb8f
Signed-off-by: Eitan Raviv <eraviv@redhat.com>