storage: remove metadata size property
In previous patch usage of size was replaced by capacity. Here we remove
size property in metadata and as a last step we make SIZE private in the
next patch - it's still needed when reading/writing legacy metadata.

This patch also removes tests which are not needed after removing size
property.

Change-Id: I6a5ef9e377ca7afb33ee7369ae920ca129f72cad
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>