nwfilter: Update vdsm-no-mac-spoofing
If the filter vdsm-no-mac-spoofing already exists, it is
updated instead of undefined and created again.
The update seems to be quick, if the content of the filter
is not changed.
This way there is no need to undefine the old filter,
which might take a long time and fail if the filter is used.

Change-Id: I5ee62e51737cda819901596097e98e315e6bb356
Bug-Url: https://bugzilla.redhat.com/1722854
Related-To: https://bugzilla.redhat.com/1720482
Signed-off-by: Dominik Holler <dholler@redhat.com>