sd: Replace alignment with max_hosts
When creating a storage domain, the user specify the block size and the
maximum number of hosts that will access the domain. Based on the number
of hosts and the block size, we select an alignment for sanlock
lockspace and resources.

Previously we converted max_hosts to alignment in
HSM.createStorageDomain(), but this conversion must be done after
detecting the underlying storage block size, which should be done in
StorageDomain.create().

The default max_hosts value was 2000. This value is very bad when using
block_size=4096 since it translates to 8 MiB alignment. The default was
changed to 250, translating to 1 MiB alignment in with any block size.

This change allows implementing block size detection in
StorageDomain.create().

Change-Id: I09ffe0f47c75cdd8efcd99e0f42fa1bc3af450af
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>