sd: Remove alignment validation
The alignment is computed using block_size and max_hosts, so there is
no need to validate the value. This was always true, but it was hidden
when alignment was computed in HSM.createStorageDomain().

Removing alignment validation, it is clear now that we must validate the
block size before computing the alignment.

Change-Id: I15102c60cc6876cb28b0c1297d86e9cd838b5e63
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>