test: Move 'hooks._scriptsPerDir' tests to 'pytest'
In an effort to move away from homegrown 'testlib' and deprecated
'nose' testing framework, we're rewriting our tests to 'pytest'
and running them with 'tox'.

This patch replaces the test from 'TestHooks' suite that validates
the 'hooks._scriptsPerDir' function with more comprehensive test set.

Some of the new tests are mark with 'xfail' to show undesired behavior
of the function:

- accepting absolute paths to hook directories just to satisfy the test
  suite itself
- allowing escaping the hook dir using "../" in path
- listing directories with executable rights as scripts
- invalid path concatenation with the "+" operator

Fixes for these issues will be introduced in next patches.

Change-Id: I33a0685129580e0c2e20419e342231106e4a8f5e
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>