sd: Detect and validate storage block size
When creating a file storage domain use the __DIRECT_IO_TEST__ file for
detecting the underlying storage block size. If the user specified
block_size=0, use storage block size. Otherwise validate that the user
specified block size matches the storage block size.

On filesystems such as NFS we cannot detect the block size because the
filesystem does not enforce minimal block size when using direct I/O.
In this case we use the user provided value (e.g. 4096), or fallback to
safe default (512).

When creating a block storage domain get the VG logical block size and
validate that it matches the user specified value.

When creating a storage domain instance, validate that the storage block
size matches the metadata block size. This is important in case the
metadata is broken or the storage was modified behind our back. This
validation is performed only if we can detect the storage block size.

Detecting block size is enabled only for localfs domain, but storage
block size validation is enabled for all domains. We will enable block
size detection gradually.

Change-Id: Ia1a2cad846db78361f2e0c034b1df96140723f67
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>