lvm: Avoid pointless retries when extending lv
In every OST run we have 80 extend LV calls, and 20 of them log this
warning:

2019-07-12 15:02:41,748-0400 WARN  (tasks/8) [storage.LVM] Command with
specific filter failed, retrying with a wider filter, cmd=['/sbin/lvm',
'lvextend', ...] rc=5 err=['  New size given (7 extents) not larger than
existing size (8 extents)'] (lvm:345)

This is the flow:

1. run lvextend to extend the lv with specifc lvm filter
2. lvextend fail because the lv is already large enough
3. retry mechanism assumes that the filter is stale, and invalidated the
   filter
4. retry mechanism run lvextend again with a wider filter
5. lvextend fail again because the lv is already large enough
6. check lv size and log a debug log that the lv is already large enough

This flow is not new, but the pointless retries were hidden in the past
since the retry mechanism was silent. Since 4.3 retries after failures
log warnings.

Most of this flow is pointless. Reduce the flow to:

1. check lv size before extending and log debug log if extend is not
   needed
2. run lvextend to extend the lv with specific lvm filter

Now the command is expected to fail only if the vg does not have enough
free extents. In this case invalidating the filter and retrying makes
sense.

Change-Id: I5e772a0ad69bdc456fd761244d6fd68feb1f3c62
Signed-off-by: Nir Soffer <nsoffer@redhat.com>