test: Move 'hooks._runHooksDir' tests to 'pytest'
In an effort to move away from homegrown 'testlib' and deprecated
'nose' testing framework, we're rewriting our tests to 'pytest'
and running them with 'tox'.

This patch replaces the test from 'TestHooks' suite that validates
the 'hooks._runHooksDir' function with more comprehensive test set.

Some of the new tests are marked with 'xfail'. There are
two reasons behind that - py3-incompatibilities and undesired
behaviors of the function exposed by the tests:

- passing unicode strings to 'os.write'
- py3-incompatible usage of 'pkgutil'
- error messages from failing scripts are lost
- improper handling of invalid utf-8 sequences
- usage of deprecated 'execCmd' function

Fixes for these issues will be introduced in next patches.

Change-Id: Ia843b3b6fa869caf1de303c4c4abd9b9e43b5dc6
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>