glusterSD: Add configuration to enable 4K support
To enable 4k support for gluster, create a drop-in configuration file:

$ cat /etc/vdsm/vdsm.conf.d/gluster.conf
[gluster]
enable_4k_storage = true

And restart vdsm to load the new configuration.

When the configuration is enabled storage block size is detected when
creating a storage domain. When the configuration is disabled,
storage with 4k sector cannot be used.

The configuration is disabled since we found issues with qemu and
qemu-img when using 4k storage. With this configuration we can backport
this to 4.3 for testing with oVirt hyperconverge setup.

Change-Id: I1835c2a21574620cfc9a24a5c51c5a74fc68bfcb
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>