docker: Add Makefile
Building images in docker is too slow and flaky lately. Add Makefile
building and pushing the images using podman.

The images can be built by anyone, but pushing will work only for people
in the vdsm[1] group in ovirtorg docker user.

We may want to add a sub stage to build the images when the build files
or docker files changes, but it is not clear how we can push the images
from jenkins to docker.

[1] https://cloud.docker.com/u/ovirtorg/orgs/ovirtorg/teams/vdsm/users

Change-Id: Iaca4b22741921ea83f9b742be53c055571f93215
Signed-off-by: Nir Soffer <nsoffer@redhat.com>