docker: Remove Fedora 28 and rawhide images
Fedora 28 is EOL for a while and the image cannot be built because of
missing packages.

Fedora rawhide image fail to build for more than a year because of
missing packages. We may add it later if someone will have time to
maintain it, and all packages will be available.

Change-Id: Id1cfb8380dd874bddad771673b85cfa495a944a6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>