net, py3: add unicode input compatibility to_str
The function to_str was not taking in account the possibility
where the input is py2 unicode, being needed to convert to string
(py2 bytes). This patch adds the encoding in case the input
is py2 unicode.

This fix is available in six [1], but is not available for rhel7,
thus we need to implement it ourselves.

[1] https://github.com/benjaminp/six/blob/master/six.py#L873

Change-Id: If6986e4fa03652b1aa2647a6cce9f0a5d44d580b
Signed-off-by: Bell Levin <blevin@redhat.com>