nmstate: bypass OVS network data validation
When the backend is nmstate, the bonds appear as not owned by
ifcfg, since they are owned by network manager. In the netswitch
configurator, those bonds are incorrectly classified as owned by
OVS.

Until we devise a mechanism to indicate a bond is owned by
NetworkManager, and that criteria is used instead of the Ifcfg
owned ones, this proposed work-around should be used, bypassing
OVS switch validation when nmstate is the backend.

Change-Id: I483214572d4ae01b10a5711c090163b938b5780f
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>