py3: virt: Don't crash on a negative socket fileno
When calling `fileno' on an inactive socket in Python 2, socket.error
is raised.  We catch that exception in GuestAgent.stop.

However in Python 3, `fileno' doesn't raise and returns -1 instead.
Vdsm crashes when it tries to use -1 file descriptor number later.
This patch avoids the problem by checking the returned file descriptor
number.

Change-Id: I50742d055780742e36a2bf96c93d3fbe9988a511
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>