nmstate: do not overwrite bond state on network deletes
When deleting a network, having a bond interface connected to
it, the bond state was being overwritten, thus preventing the
bond from being deleted.

By not overwritting the bond state configuration, we can keep
having a clear separation between the network and bond entities
when deleting the entities - e.g. the bond entity is responsible
for creating/deleting the bond, whereas the network entity is
responsible for using it.

Change-Id: I57e352c4d0629cd20c873054f36dc3cdeb3591b4
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>