persistent: Do not log binary data
When using gluster/vdo we get full block padded with zeros. The last
line full of zeros is not valid metadata, so we log it. Here is example
log line:

2019-08-01 23:02:17,001+0300 WARN  (monitor/de56647)
[storage.PersistentDict] Could not parse line `^@^@^@...

Logging the string as is breaks grep and less later, since the log
contains binary data.

Use "%r" format to log a safe representation of the possible binary
data, converting this log to:

2019-08-01 23:02:17,001+0300 WARN  (monitor/de56647)
[storage.PersistentDict] Could not parse line `\x00\x00\x00...

Change-Id: I9107344c4b1371278a8f4e3ec9662a70fd228933
Related-To: https://bugzilla.redhat.com/1737141
Signed-off-by: Nir Soffer <nsoffer@redhat.com>