travis: Remove Fedora 28 build
Fedora 28 is EOL for a while, but we kept this build because we don't
have a CentOS 8 image with python 3.6, and this was the last distro with
python 3.6.

This builds tends to break every few weeks, and limit our options (e.g.
missing xfsprogs). Lets kill it.

Also remove stale fc28 file from the automation directory, that was
forgotten when we removed the automation for fc28.

Change-Id: Iaae05c704fcd9ab7f9cd858daae6f77a92c5dba4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>