nmstate: fix the current configuration update
Before this patch, we were storing in the running configuration
the networks - and bonds - of the current 'setupNetworks'
instruction, thus overwriting its current contents.

With this change, we add/remove the new networks / bonds to the
current configuration.

To preverse current behavior, in which a setupNetworks instruction
having both a remove and add/edit instruction *keeps* the network
defined, we first process the 'remove' instructions, and only then
the 'add' instructions.

Change-Id: I1ffb524e24d1d7f3459238075a1600138d0ba32a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>