storage.compat: Support wait argument in sanlock for python 2
Sanlock for python 3 changed add_lockspace and rem_lockspace to use
wait=True instead of async=False (which is a keywoard in python 3.7).

Wrap the python 2 functions with decorator converting wait=True to
async=False, so we can update the application code to use the new style,
while keeping compatibility with python 2 for testing.

The decorator should be removed when we drop support for python 2.

Change-Id: I666c11593543f01916b79075b614b829a002d069
Signed-off-by: Nir Soffer <nsoffer@redhat.com>