net: Add dhclient as default dhcp for NM controlled nics
Currently vdsm decides whether the interface has dhcp configuration
or not by looking into running proccesses searching for dhclient.

If the dhclient is running with particular flags e.g. -6
vdsm fills this as 'dhcpv4' or 'dhcpv6' capabilites of the interface.

NetworkManager by default uses it's internal dhcp client which is not
running as separate process. Add configuration that forces NetworkManager
to use dhclient as it's dhcp client.

Change-Id: I531fb8122b705633faafd32048a5c5d84bfc71e1
Bug-Url: https://bugzilla.redhat.com/1738423
Signed-off-by: Ales Musil <amusil@redhat.com>