net func tests: Do not use Bridges with no ports
oVirt does not support portless bridges, however there is a test that
uses portless bridges.

The `test_create_network_over_an_existing_unowned_bridge` is such a
test. This change adds to the test bridge a dummy interface as a port.

This has worked fine with the initscripts as the backend, however, with
Nmstate/NM as the backend, portless bridgess are set as DOWN, causing
the functional tests assertions to fail.

Change-Id: I2297a8d0ebb766df9427d76fcb6951b127e7c20e
Signed-off-by: Edward Haas <edwardh@redhat.com>