py3: tests: 'rpc.http.ImageRequestHandler' tests
This patch adds tests for 'rpc.http.ImageRequestHandler' class to help
us in making it py3-compatible. The tests cover class' codebase
almost in 100%.

Due to the architecture of the RPC facilities of VDSM the tests
need to include an implementation of a simple HTTP server.

Since the tests don't work with py3 yet, they've been marked with
'skipif' decorator. Choosing 'skipif' over 'xfail' is mandatory here -
when 'xfail' is used the tests are still being run and they reach
a timeout due to some events not being set.

Change-Id: I7404e2a186c5360291bfaa39a7f35e83a9a6dc13
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>