nmstate, net: Get dhcp info from nmstate
Get dhcp info from nmstate backend to *fix* VDSM reporting about
`autoconf` which it wrongly detects as enabled when NM is
managing the interface. As a subsequence of this, when no dynamic
configuration exists on an ipv6 interface, VDSM reports back
autoconf=True and as a result Engine will issue a setupNetworks with
dhcpv6=False and autoconf=True, which in turn is rejected by
nmstate/NM as being a not supporting scenario.

Change-Id: I732ea2afced4d892fd3c00a70e8ab6062c750e45
Signed-off-by: Ales Musil <amusil@redhat.com>