net: Set NM dhcpv4 request ID to the interface mac
In scenarios where a bridged network is added over an existing NIC with
a dynamic IP, the IP is expected to migrate to the bridge and kept
dynamic. This has been working fine with the initscripts but it is not
the case with NM managing all interfaces.
NM default in this regard needs to be updated such that by default, the
dhcpv4 request will use the interface mac address as its identifier.

Change-Id: Iba3b2dd6d4862fc65e6a2a96cdf33a74fce3a384
Signed-off-by: Edward Haas <edwardh@redhat.com>