py3: mkimage: Remove md5 part in getFileName
We calculate checksum in 'mkimage.getFileName' and it is apparently 
triggered several times with presumably different content because the
hash differs. Probably only whitespace differences since ultimately
the content is actually correct.
However, we do not really need a checksum as engine always sends only 
one payload. In this scenario a file with a name derived only from
the VM id is sufficient, as it's guaranteed that there will be only one
created.

Change-Id: I38b4c6107038bb8129036411ba26045ff1fc949c
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>