mom client: Add timeout to the RPC connection
In case MOM hangs, the RPC requests would never finish and
the VDSM worker would be blocked. MOM is queried periodically,
so after a while, all VDSM workers would be blocked.

This patch adds timeout to the connection, so workers will
not be blocked forever.

Another change is that MOM client remembers if a previous request
failed or timed out and does not send another request for a certain time.
This is to avoid multiple successive requests waiting to time out.

Change-Id: Ib5bf49743c00557dbecc052b3a18dd3dae8083e2
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>