py3: sslutils: Adapt 'SSLSocket' class to py3
We maintain our own 'SSLSocket' class in 'sslutils' module. The reason
behind this is our protocol detection functionality - to make it work,
we need to peek inside the incoming data and pass the socket to next
layers based on first few bytes. The bytes need to stay there though -
we can't remove them from the buffer. Unfortunately, the 'ssl.SSLSocket'
class that comes with the standard library doesn't support
the 'MSG_PEEK' flag (or, as a matter of fact, any other flag) [1] -
hence our own socket implementation.

To make our own 'SSLSocket' work with py3 we need two changes.

The first one is about replacing a call to non-existent in py3
'socket._fileobject' method. It's being replaced with a call
to 'socket.socket.makefile' method, which for py2 simply forwards
the call to '_fileobject' [2].

The second change is about providing missing 'recv_into' implementation.
This one is needed because of the changes in implementation
of 'http.server' in py3. Our own image request handler HTTP server
is based on the one that comes from the standard library [3].
When the server handles a request it uses the 'readline' method
on the socket, which calls 'socket.socket.readinto' method,
which leads to 'socket.socket.recv_into' call [4]. Unfortunately
'recv_into' is not using 'recv' under the hood - it uses internal,
C-implemented 'sock_recv_guts' [5]. Our 'recv_into' implementation
is based on the implementation of this method for the 'socket.socket'
class, the 'ssl.SSLSocket' one [6] and the docs [7].

[1] https://docs.python.org/3.6/library/ssl.html?highlight=ssl%20sslsocket#ssl.SSLSocket
[2] https://github.com/python/cpython/blob/0229b56d8c0cb65b8ad789e69dcd281fd92a6d96/Lib/socket.py#L221
[3] https://github.com/oVirt/vdsm/blob/e30ef3a05c8286fab0714a198833c33e73d9ae0c/lib/vdsm/rpc/http.py#L271
[4] https://github.com/python/cpython/blob/f1f9c0c532089824791cfc18e6d6f29e1cd62596/Lib/socket.py#L586
[5] https://github.com/python/cpython/blob/f1f9c0c532089824791cfc18e6d6f29e1cd62596/Modules/socketmodule.c#L3197
[6] https://github.com/python/cpython/blob/f1f9c0c532089824791cfc18e6d6f29e1cd62596/Lib/ssl.py#L1001
[7] https://docs.python.org/3.6/library/socket.html?highlight=socket%20recv_into#socket.socket.recv_into

Change-Id: Ia5ea7c649b4b09690afb5c8ad5b46866431a40c2
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>