tests, net, nmstate: add 2 networks with dynamic IP
When using dynamic IP networks, routes and nameservers can be learnt
from the network. If we're using multiple networks, there is a
possibility that we learn these routes and nameservers from both
networks.

This test ensures that the routes and gateways are only learnt via
dhcp for the *default route*.

Change-Id: I07aa43b04b8214c7f5461b442e1c7a20fb61e31c
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>