virt: Add support for TLS migrations
TLS migrations are enabled by VIR_MIGRATE_TLS migration flag.  It's
added when Engine requests an encrypted migration via `encrypted'
migration parameter.

In order to successfully migrate over TLS, we need to match the
destination host name with the destination host certificate.  Vdsm
currently uses an IP address to connect to the destination libvirt,
which wouldn't match.  To fix that, we use a host name, rather than an
IP address, to connect to the destination libvirt when an encrypted
migration is requested.  This may break on certain setups, e.g. when a
separate migration network is used or when IPv4/IPv6 dual stack is
used, but that's currently an inherent limitation when encrypted
migrations are used.  The solution is to add a libvirt migration
parameter to check the destination certificate against a given host
name; however this is not yet implemented in libvirt, see
https://bugzilla.redhat.com/1754533.  Until the feature is available,
we have to use the workaround described above.

Change-Id: I6e38c9f2802c579b76668ce6772c10866983996a
Bug-Url: https://bugzilla.redhat.com/1739557
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>