storage: introduce transientdisk module
The transientdisk module created in order to provide an API for adding
a temporary disks on a temporary directory.

The module will be responsible to create the transient disks directory under -
/var/lib/vdsm/storage/transient_disks which creates when VDSM starts,
when all the transient disks will be removed from the directory,
the directory will be deleted.

This module is needed for creating "scratch" disks for the full VM backup
which is the first step in the incremental backup flow.

The creation and removal of the transient disk added as public API for
future usage, currently the API is used internally only.

Change-Id: If01c59b9185fa4eb7ce4be247605563db41303a2
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>