tests: Re-enable py-watch's backtrace test
We run our tests with 'tests/py-watch' wrapper to handle timeouts.
It has a nice feature - it tries to dump a python call stack of the
hanged process before killing it. It's possible by attaching with gdb
and calling 'thread apply all py-bt'. To make this feature work, we need
python debuginfo package installed that matches exactly the version
of the python package itself.

Since debuginfo packages are kept in separate repos, it may happen,
that they go out of sync with the proper package. This patch will make
us aware of the situation by:

- printing a warning that the versions don't match just after installing
  the debuginfo package in 'automation/common.sh'
- marking 'test_timeout_backtrace' test as expected to fail

The timeout backtrace test has been disabled for quite some time and it
seems that the output of 'py-bt' command has changed slightly, example
output:

 (gdb) py-bt
 Traceback (most recent call first):
   <built-in method sleep of module object at remote 0x7fe9e61c4fb0>
   File "sss.py", line 8, in inner
     time.sleep(100000)
   File "sss.py", line 5, in outer
     inner()
   File "sss.py", line 10, in <module>
     outer()

Assertions have been aligned to current output.

Change-Id: I4e8e211702cf3844f366fba7287b25f0c7ad3dca
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>