tests, net: validate network exists upon deletion
The validation of the existence of the bridge was only checked
when the network backend were ifcfg files.

This patch moves this validation higher in the stack, and as a
result, the nmstate backend can also benefit from it.

Two unit tests are added, one ensuring that a BAD_BRIDGE error
code is thrown when attempting to delete a network that does not
exist, and another ensuring that a 'broken' network (a network
currently in the RunningConfig but not reported by the kernel) does
not thrown the BAD_BRIDGE error.

Change-Id: I0e4ce0ee9aec206f3d8c483eef9e8e1173337d89
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>