tests: add test_copy_data_collapse
This patch introduces a test to prove the current
SDM.copy_data implementation can support collapse properly,
as this is needed to support flows like VM cloning.

The patch creates a chain of 3 images, writes some data
to each layer, copies the chain to a single target volume
and then verifies the data written to the original chain
is present on the target.

Change-Id: Ia5c21619abb51e906642086103e5c89f8725ab4d
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>