virt: Don't fail on payload removal in releaseVm
When removal of a payload fails in releaseVm, an exception is raised
and the whole operation fails.  That prevents Engine from cleaning up
the VM and the VM may be shown as running and non-destroyable in
Engine even when it is already down.

In theory, it shouldn't happen, but due to various bugs it can happen,
e.g. when the payload file is missing for some reason.  For instance,
when a wrong payload file path is provided, libvirt may remove the
drive from the domain XML, but Vdsm still expects the file to be
present and fails in releaseVm as described above.

Let's not fail when the payload file can't be removed, just log it.
The worst what can happen is that the payload file remains present
until (if ever) it gets replaced.  If it is non-removable, the given
VM may be prevented from starting with a payload file on the given
host, but this situation is very unlikely and needs an administrator's
action in any case.  That all is better than not being to destroy an
already stopped VM.

Change-Id: I6d60c4a1f5fd2215e6948bf565a6a3aa84c27a49
Bug-Url: https://bugzilla.redhat.com/1752866
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>