virt: Use the provided payload path on incoming migrations
Commit cbe05c00b36186c935b5808d4cb00b74c126f87a has changed naming of
payload files.  That means the payload file has different names in
different Vdsm versions and thus its original path given in the domain
XML may no longer be applicable when the VM is migrated to another
host running a different Vdsm version.  libvirt removes the
corresponding drive from the VM in such a case and the payload gets
inaccessible to the VM.

To remedy this problem, let's use the original payload path to
generate a new copy of the payload file on incoming migrations.  This
will solve the issue for VMs migrated from hosts running older Vdsm
versions and thus unaware of this problem.  A followup patch will
address the problem when migrating to hosts running older Vdsm
versions.

Change-Id: Iac0d0be5d1d9391c6b533c99d7f3d25458785c14
Bug-Url: https://bugzilla.redhat.com/1752866
Backport-To: 4.3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>