virt: Change the payload path for outgoing migrations
The payload path has been changed in Vdsm 4.4.  In order to support
migrations of VMs with payloads to hosts running older Vdsm versions,
we must change the payload path to its old value before sending the
domain XML to the destination host.

We cannot simply make the legacy payload path in the old way for two
reasons:

- The legacy path relayed on Python 2 dictionary ordering.

- It was constructed using md5 sums, which are better to avoid in
  order to not confuse security checkers (although no security is
  involved in the path construction).

So the easiest, although quite ugly, way to get the legacy payload
path is to ask the destination Vdsm for its payload path.  Then the
updated path is put into the domain XML that is sent to the
destination host as a migration parameter.

Unfortunately we can't handle hibernation in a similar way, since we
don't know in advance the host where the VM will be restored.

Change-Id: I4fe2b21f3b2cd0fbc143d2938d38bbf625adff9c
Bug-Url: https://bugzilla.redhat.com/1752866
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>