spec: Require qemu-kvm supporting 4k for RHEL 7.7.z
To enable 4k support with Gluster storage, we need qemu-kvm version that
can probe alignment correctly. This version is available for testing in
QE repos, and will be released in next 7.7.z update.

We don't enable 4k storage by default since Gluster QE did not finish
testing 4k. We plan to enable 4k storage for gluster in 4.3.8.

Note: This is 4.3 only patch, on master we don't support RHEL 7.7.

Change-Id: Ie9fb092e447f18d8f67dc1a48b582461d1f5eb04
Bug-Url: https://bugzilla.redhat.com/1748022
Related-to: https://bugzilla.redhat.com/1745443
Signed-off-by: Nir Soffer <nsoffer@redhat.com>