travis: Fix Fedora 30 build
On travis lvm tests fail with:

    taskset: failed to execute /sbin/lvm: No such file or directory

Adding the lvm2 package to the Fedora 30 dockerfile fixes the lvm tests.

This change also fixes creating loop devices, so now all tests pass on
Fedora 30, and we don't need to allow failures.

We keep the Fedora 29 python 3.7 tests for a while in case the Fedora 30
build is not stable.

Change-Id: I0c90de74de1afe72b971cbf2b493bdf4710f269b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>