tests: Make xfail strict by default
Tests marks as xfail do not fail the build if the test pass. Instead,
they show as XPASS. XPASS tests are bad since they may fail silently,
hiding regressions after the original issue was fixed.

Our experience shows that we are not good at removing stale xfails, so
we are moving to strict by default. Now when fixing a bug, stale xfails
will break the build instead of passing as XPASS.

To handle flaky tests that fail randomly, you can override the strict
argument on the flaky test:

    @pytest.mark.xfail(strict=False)
    def test_flaky_code():
        ...

Change-Id: Ie98f6dc921385579025144fa43ab2c5169c8b81f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>