tests: Fix debuginfo version check on CentOS
On CentOS 7, we don't have python2 package, so debuginfo-install fails
to install python2-debuginfo:

    No debuginfo packages available to install

(The build does not fail since we use -y)

And then we log a warning:

    WARNING: python2-debuginfo- doesn't match python2-! Backtraces for
    timed-out tests won't be available!

It is true that python2-debuginfo is not installed, but we can see that
python2 version is incorrect, since "rpm -qa" also does not know about
"python2" package.

I think this issue was introduced by starting to use CI_PYTHON for
debuginfo-install. The issue was hidden since the test was marked as
xfail, and we did not use strict xfail.

Fix the issue by falling back to "python" and "python-debuginfo" if
CI_PYTHON is "python2", and we fail to query "python2" version info.

With this change pywatch backtrace test passes on all platforms.

Change-Id: If69689c856787de04cf7beeb1433ecb706b3d01b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>