task: Fix race when aborting tasks
When aborting a task, if the task job completes in the task thread
before the aborting thread called _doAbort, we move into "raborting"
state, and then fail to start recovery with this error:

    TaskStateTransitionError: cannot move task to requested state:
    ('from aborting to raborting',)

The race is:

[aborting thread] set self._aborting to True
[aborting thread] invoke the abort callback
[task thread] wakes up and task Job completes
[task thread] exit normally and change state to "raborting"
[aborting thread] invoke _doAbort
[aborting thread] fail to change state from "aborting" to "raborting"

The current code checked if a task was aborted before running a Job,
so if we have multiple jobs, we would detect that the task was aborted
when running the next job. However if we have only one job, we continue
normally instead of aborting.

Fix by checking if task was aborted also after a job completes, and
raising TaskAborted.

Change-Id: Ib11a2ee8a482c3a70f0b694280859b9cdc5a4e67
Bug-Url: https://bugzilla.redhat.com/1751881
Signed-off-by: Nir Soffer <nsoffer@redhat.com>