tests, storage: Add Callable class to storagetestlib
This will serve following patches for Task tests.

Callable has 'hang' mode (set with hang=True on creation)
allowing it to run its callable part indefinitely until
finish() is called. This is useful for async executions where
we want to control Task duration. We also provide timeouts
for the waitful parts so in case something goes wrong and hangs
for too long in the MainThread, a RuntimeError for failing
threading.Event.wait() will be raised.

Change-Id: I47f2f9267e162d2861f0a73b9b8f32ac0733b8ad
Signed-off-by: Amit Bawer <abawer@redhat.com>