storage: Add wait parameter to taskManager.prepareForShutdown()
If 'wait' is set to False, underlying threadPool will not wait for its worker
threads to finish before shutting down, this is the default and the behavior so far.

If 'wait' is set to True, threadPool will await for its threads to finish before
shutting down.

Using wait=True will prevent from leaking threadPool threads in tests.

Change-Id: Icdf33633ee1a5bec417203ff5ce3aa6203d6d62c
Signed-off-by: Amit Bawer <abawer@redhat.com>