sd: Add getVolumeSize()
getVolumeSize() replaces getVSize() and getVAllocSize() which are
deprecated and should not be used in new code.

On block storage, getVSize and getVAllocSize are the same method, so
when invoking HSM.getVolumeSize() we call getVSize() twice.

On file storage both getVSize and getVAllocSize are implemented using
oop.os.stat(). When invoking HSM.getVolumeSize() we call oop.os.stat()
twice. Every oop call may timeout after 60 seconds, so in the worst case
this can cause a delay of 120 seconds.

In both storage types if accessing storage is very slow, the caller must
pay twice for the slow access for no reason. With new getVolumeSize() we
access storage only once for getting both apparent and true size.

This change cannot fix the root cause for slow getVolumeSize() calls,
but in some cases it may decrease the delay by half.

Change-Id: Ic9edd752bc5e87c96fc2dc400b5f738647623c93
Related-to: https://bugzilla.redhat.com/1598266
Signed-off-by: Nir Soffer <nsoffer@redhat.com>